Implement --factor option
